+++ src/schemathesis/_hypothesis.py
@@ -4,6 +4,7 @@ import re
 from base64 import b64encode
 from functools import partial
 from typing import Any, Callable, Dict, Generator, Optional
+from urllib.parse import quote_plus
 
 import hypothesis
 import hypothesis.strategies as st
@@ -112,7 +113,7 @@ def get_case_strategy(endpoint: Endpoint) -> st.SearchStrategy:
     static_kwargs = {"path": endpoint.path, "method": endpoint.method, "base_url": endpoint.base_url}
     try:
         strategies = {
-            "path_parameters": from_schema(endpoint.path_parameters),
+            "path_parameters": from_schema(endpoint.path_parameters).filter(filter_path_parameters).map(quote_all),
             "headers": from_schema(endpoint.headers).filter(is_valid_header),  # type: ignore
             "cookies": from_schema(endpoint.cookies),
             "query": from_schema(endpoint.query),
@@ -121,6 +122,19 @@ def get_case_strategy(endpoint: Endpoint) -> st.SearchStrategy:
         return _get_case_strategy(endpoint, static_kwargs, strategies)
     except AssertionError:
         raise InvalidSchema("Invalid schema for this endpoint")
+
+
+def filter_path_parameters(parameters: Dict[str, Any]) -> bool:
+    """Single "." chars are excluded from path by urllib3.
+
+    In this case one variable in the path template will be empty, which will lead to 404 in most of the cases.
+    Because of it this case doesn't bring much value and might lead to false positives results of Schemathesis runs.
+    """
+    return not any(value == "." for value in parameters.values())
+
+
+def quote_all(parameters: Dict[str, Any]) -> Dict[str, Any]:
+    return {key: quote_plus(value) if isinstance(value, str) else value for key, value in parameters.items()}
 
 
 def _get_case_strategy(