+++ docs/changelog.rst
@@ -16,6 +16,11 @@ Added
 - Pre-run hooks for CLI. `#147`_
 - A way to register custom checks for CLI via ``schemathesis.register_check``. `#270`_
 
+Fixed
+~~~~~
+
+- Not encoded path parameters. `#272`_
+
 Changed
 ~~~~~~~
 
@@ -393,6 +398,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#272: https://github.com/kiwicom/schemathesis/issues/272
 .. _#270: https://github.com/kiwicom/schemathesis/issues/270
 .. _#268: https://github.com/kiwicom/schemathesis/issues/268
 .. _#261: https://github.com/kiwicom/schemathesis/issues/261